Move Action & GeneralAction to actions package
