+++ domain/src/main/java/com/aconno/sensorics/domain/ifttt/ActionsRepository.kt
@@ -1,5 +1,6 @@
 package com.aconno.sensorics.domain.ifttt
 
+import com.aconno.sensorics.domain.actions.Action
 import io.reactivex.Single
 
 /**
