+++ domain/src/main/java/com/aconno/sensorics/domain/interactor/ifttt/InputToOutcomesUseCase.kt
@@ -1,6 +1,6 @@
 package com.aconno.sensorics.domain.interactor.ifttt
 
-import com.aconno.sensorics.domain.ifttt.Action
+import com.aconno.sensorics.domain.actions.Action
 import com.aconno.sensorics.domain.ifttt.ActionsRepository
 import com.aconno.sensorics.domain.ifttt.Input
 import com.aconno.sensorics.domain.ifttt.outcome.Outcome
